Fixes segmentation fault in `picoquic_incoming_encrypted`
When the certificate of a client can not be verified, but the client already
send an encrypted package, the server dies with a segmentation. In the
`picoquic_incoming_encrypted` function, the state is checked for `>=
client_almost_ready` and this returned true. With this reordering, the check
does not pass anymore and the server does not crash.