appveyor: allow Debug|x64 to fail due to linker issue
For some reason AppVeyor (VS2017 14.15.26726) fails to link
picoquicdemo due to an issue related to JMC[1]. Apparently JMC requires
an application to be linked with a CRT which somehow does not seem to
happen. JMC can be disabled through "/JMC-" as described at [1].
Ignore the issue until a proper solution is found.

 [1]: https://blogs.msdn.microsoft.com/vcblog/2018/06/29/announcing-jmc-stepping-in-visual-studio/