Set `stop_sending_sent` even if we don't need to send it
The calling function checks if `stop_sending` was requested and it was
not sent yet. Together with the new `is_active` feature this can result
in an endless loop. `stop_sending_sent` is not set to 1, but `is_active`
stays at 1, the stream does not fill the send_buffer and
`picoquic_find_next_ready_stream` always returns the same stream.