WIP - create a binary log file per connection
Currently picoquic writes one giant binary log file that contains all traffic that picoquic handles across all connections. The advantage is that even context-free information can be written and later matched to the correct connection once the connection id is known. The disadvantage is that it is hard to implement log-file rotation and it is more difficult to extract connection data from this log file when a connection is closed.

This change allows picoquic to be configured with a directory where log files are written per connection.