Fix parsing of coalesced packets
Is very unlikely that a coalesced packet with an unexpected version
is a valid QUIC packet. It is probably some kind of unencrypted random
padding data added after a previously valid QUIC packet.

Note that we don't perform any checks on CIDs presence/values for any
packets with an unsupported version; if the coalesced packet is very
short , we might read invalid data when preparing the VN reply.